+++ WrightTools/_group.py
@@ -340,18 +340,16 @@ class Group(h5py.Group, metaclass=MetaClass):
         if name is None:
             name = self.natural_name
         if parent is None:
-            from ._open import open as wt_open  # circular import
-
             new = Group()  # root of new tempfile
             # attrs
             new.attrs.update(self.attrs)
-            new.natural_name = name
             # children
             for k, v in self.items():
                 super().copy(v, new, name=v.natural_name)
             new.flush()
-            p = new.filepath
-            new = wt_open(p)
+            # Converts to appropriate Data/Collection object
+            new = new["/"]
+            new.natural_name = name
         else:
             # copy
             self.file.copy(self.name, parent, name=name)