+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Added
+- `artists.create_figure`: kwarg `margin` allows unique margins on all sides
+
 ### Fixed
 - ZeroDivisionError when converting units (e.g. wn to nm) now returns inf instead of raising
+- PermissionError on windows when copying Data/Collection objects
 - Fixed bug in `wt.artists.interact2D` where constants were not handled properly.
 
 ## [3.4.0]